Successfully sorted out #imported-modules and #modules
This was done by looking at guix/build-system/cmake.scm

They define a function cmake-build which is the (builder) field in the bag.
That's what gets passed the #:imported-modules and the #:modules.

The #:imported-modules are used with the with-imported-module form outside the build g-exp:
This is essentially "making the modules available.

The #:modules are used with (use-modules) directly.  So these are the actual
module imports that are used in the gexp.

The final trick is that we #:select the symbols we need to avoid conflicting symbols.
In particular, cmake-build-system is needed because it defines cmake-build (the function in cmake-build-system that actually does the building. However it also exports %standard-phases,
but we want %standard-phases from catkin-build-system.