Added pluginlib.
Also needed to change some propagated inputs to inputs.
It's not always clear from package.xml what type of inputs it should be (or I'm
reading it wrong).

Here's what I'm doing:
If I have as inputs all the dependencies listed in a package.xml
and the package doesn't build because of a missing dependency,
I then find the 2nd order dependencies that have the missing one and make
a propagated input.