roscore runs in shell environment!
Working on a pattern of setting paths and modifying executables
to use them (without exposing the whole system).

The pattern is as follows:
Create a new search-path-specification with the path required.
Make that a native-search-path for the package in question.

Wrap the executables, set the actual variable in question to the variable.

e.g.,

set NOETIC_ROS_CMAKE_PREFIX_PATH to the ROS_PREFIX directory (basically the root of the profile)
then wrap each catkin executable, and set CMAKE_PREFIX_PATH=${NOETIC_ROS_CMAKE_PREFIX_PATH}