Made the chdir done with #:package-dir be part of the unpack phase
Before it was added after 'unpack. But then if someone else
added after 'unpack, what would happen?

The idea of changing the directory is that the code you want for a project
is under a specific subdirectory, so the result of unpack should
be that you are in that project's directory with all the necessary code.