Adding qt core.
This required changing a dependency on python-rospkg

python-pyrpoject is propagated by pydot and rospkg,
but pydot depends on pyproject-2.4.7 so we needed to change this in rospkg too.

Other options were:
1. Updating pydot
2. creating anew variant

But overall, the old version is already in guix and it is closer to
the version used on Ubuntu 20.04 so we'll just use that version for now.