Added GUIX_ROS_CMAKE_PREFIX_PATH to native search paths for roslaunch.
This allows roslaunch to find packages in the profile.

Essentially, at runtime we do not wrap all ROS projects to find all ros nodes,

instead we rely on native-search-paths to put the ros packages in the profile on the path
(otherwise every package that used a launchfile would need to specify paths to each ros package that was used,
which could result in cleaner profiles but seems like overkill).

Before catkin set this path, but if roslaunch is used on its own catkin is not in the profile so the search path is not set.